:truck: move constants folder.
