+++ packages/playground-react/constants/icons.tsx
